Math Problem Square Block (Not optimized)
In a map, there are two roads. Each road can be divided into multiple segments of integer length. Write a piece of code that finds a number that represents the largest  size of the segment in a square city block created from the two roads. If not possible, the code returns 0.
Examples:
1) Given A = 10, B = 21. The solution is 7. The second road is split into three segments of length 7 and the first is cut into one segment of length 7, and the segment of length 3 is ignored.
2) Given A = 13, B = 11. The solution is 5. You can get two segments of length 5 from each road.

Make the solution as efficient as possible.